backend/requirements.txt
fastapi
uvicorn
requests
supabase
python-jose
python-dotenv
python-multipart
